synchronize the file access in loaders
so loaders don't accidently mess around with other's cachefile.

sometimes the file is still being prefetching, but we scroll over to
that place and request for DisplayImage(). getBitmapDisk() would fail,
because the file isn't downloaded yet, but it then ask the webloader to
download it again.

by synchronizing the access to each file, we ensure there's only one
thread accessing a particular file at a point.